Add RequestLogEventBuilderFactory
