+++ server/src/main/java/org/apache/druid/server/log/EmittingRequestLogger.java
@@ -19,35 +19,30 @@
 
 package org.apache.druid.server.log;
 
-import com.fasterxml.jackson.annotation.JsonProperty;
-import com.fasterxml.jackson.annotation.JsonValue;
-import com.google.common.collect.ImmutableMap;
-import org.apache.druid.guice.annotations.PublicApi;
-import org.apache.druid.java.util.emitter.core.Event;
 import org.apache.druid.java.util.emitter.service.ServiceEmitter;
-import org.apache.druid.java.util.emitter.service.ServiceEventBuilder;
-import org.apache.druid.query.Query;
-import org.apache.druid.server.QueryStats;
 import org.apache.druid.server.RequestLogLine;
-import org.joda.time.DateTime;
-
-import java.util.Map;
 
 public class EmittingRequestLogger implements RequestLogger
 {
-  final ServiceEmitter emitter;
-  final String feed;
+  private final ServiceEmitter emitter;
+  private final String feed;
+  private final RequestLogEventBuilderFactory requestLogEventBuilderFactory;
 
-  public EmittingRequestLogger(ServiceEmitter emitter, String feed)
+  EmittingRequestLogger(
+      ServiceEmitter emitter,
+      String feed,
+      RequestLogEventBuilderFactory requestLogEventBuilderFactory
+  )
   {
     this.emitter = emitter;
     this.feed = feed;
+    this.requestLogEventBuilderFactory = requestLogEventBuilderFactory;
   }
 
   @Override
   public void log(final RequestLogLine requestLogLine)
   {
-    emitter.emit(new RequestLogEventBuilder(feed, requestLogLine));
+    emitter.emit(requestLogEventBuilderFactory.createRequestLogEventBuilder(feed, requestLogLine));
   }
 
   @Override
@@ -56,96 +51,7 @@ public class EmittingRequestLogger implements RequestLogger
     return "EmittingRequestLogger{" +
            "emitter=" + emitter +
            ", feed='" + feed + '\'' +
+           ", requestLogEventBuilderFactory=" + requestLogEventBuilderFactory +
            '}';
-  }
-
-  @PublicApi
-  public static class RequestLogEvent implements Event
-  {
-    final ImmutableMap<String, String> serviceDimensions;
-    final String feed;
-    final RequestLogLine request;
-
-    RequestLogEvent(ImmutableMap<String, String> serviceDimensions, String feed, RequestLogLine request)
-    {
-      this.serviceDimensions = serviceDimensions;
-      this.request = request;
-      this.feed = feed;
-    }
-
-    @Override
-    // override JsonValue serialization, instead use annotations
-    // to include type information for polymorphic Query objects
-    @JsonValue(value = false)
-    public Map<String, Object> toMap()
-    {
-      return ImmutableMap.of();
-    }
-
-    @Override
-    @JsonProperty("feed")
-    public String getFeed()
-    {
-      return feed;
-    }
-
-    @JsonProperty("timestamp")
-    public DateTime getCreatedTime()
-    {
-      return request.getTimestamp();
-    }
-
-    @JsonProperty("service")
-    public String getService()
-    {
-      return serviceDimensions.get("service");
-    }
-
-    @JsonProperty("host")
-    public String getHost()
-    {
-      return serviceDimensions.get("host");
-    }
-
-    @JsonProperty("query")
-    public Query getQuery()
-    {
-      return request.getQuery();
-    }
-
-    @JsonProperty("remoteAddr")
-    public String getRemoteAddr()
-    {
-      return request.getRemoteAddr();
-    }
-
-    @JsonProperty("queryStats")
-    public QueryStats getQueryStats()
-    {
-      return request.getQueryStats();
-    }
-
-  }
-
-  private static class RequestLogEventBuilder extends ServiceEventBuilder<Event>
-  {
-    private final String feed;
-    private final RequestLogLine requestLogLine;
-
-    public RequestLogEventBuilder(
-        String feed,
-        RequestLogLine requestLogLine
-    )
-    {
-      this.feed = feed;
-      this.requestLogLine = requestLogLine;
-    }
-
-
-    @Override
-    public Event build(ImmutableMap<String, String> serviceDimensions)
-    {
-      return new RequestLogEvent(serviceDimensions, feed, requestLogLine);
-    }
   }
 }
